requirements.txt
pandas
openpyxl
PyQt6
bs4
Pyyaml